[develop] | changes for fetch data in App component and declare propTypes for List component
