+++ spotifood/src/components/App.js
@@ -1,18 +1,31 @@
 // Global
-import React, { Component } from 'react';
+import React, { useEffect, useState } from 'react';
 // Components
 import List from './List';
+// Api
+import { getPlaylists } from '../api';
 // Stylesheets
 import './App.scss';
 
-class App extends Component {
-    render() {
-        return(
-            <div className='page'>
-                <List />
-            </div>
-        )
-    }
+function App() {
+    const [list, setList] = useState();
+
+    useEffect(() => {
+        async function fetchData() {
+            const response = await getPlaylists();
+            const data = response.playlists.items;
+
+            setList(data);
+        }
+
+        fetchData();
+    }, []);
+
+    return(
+        <div className='page'>
+            <List data={list} />
+        </div>
+    )
 }
 
 export default App;
